Move Notification and Notifier classes to controller package
